refactor: reorganize imports and streamline bot initialization; remove unused spam handling logic
